fix full build
